- Keep neoverse-v2 FetchUnit (mopCache, no loop buffer) over dev's loop buffer approach
- Keep neoverse-v2 branch predictor (speculative global history, FTQ) over dev's program-order verification
- Take dev's AArch64 instruction support (SME2, NEON, SVE2, new opcodes)
- Take dev's Instruction_decode refactor to AARCH64_ enum constants
- Keep both instruction implementations where HEAD/dev had non-overlapping opcodes
- Take dev's PACIASP/AUTIASP improved implementation
- Take dev's DecodeUnit back/pop_back bug fix

+++ src/include/simeng/pipeline/PortAllocator.hh
@@ -16,7 +16,7 @@ const uint8_t OPTIONAL = 1;
 /** An abstract execution port allocator interface. */
 class PortAllocator {
  public:
-  virtual ~PortAllocator(){};
+  virtual ~PortAllocator() {};
 
   /** Allocate a port for the specified instruction group; returns the allocated
    * port. */
@@ -33,7 +33,7 @@ class PortAllocator {
   /** Set function from DispatchIssueUnit to retrieve reservation
    * station sizes during execution. */
   virtual void setRSSizeGetter(
-      std::function<void(std::vector<uint64_t>&)> rsSizes) = 0;
+      std::function<void(std::vector<uint32_t>&)> rsSizes) = 0;
 
   /** Tick the port allocator to allow it to process internal tasks. */
   virtual void tick() = 0;
