+++ test/unit/pipeline/M1PortAllocatorTest.cc
@@ -8,23 +8,23 @@ class M1PortAllocatorTest : public testing::Test {
  public:
   M1PortAllocatorTest() : portAllocator(portArrangement, rsArrangement) {
     portAllocator.setRSSizeGetter(
-        [this](std::vector<uint64_t>& sizeVec) { rsSizes(sizeVec); });
+        [this](std::vector<uint32_t>& sizeVec) { rsSizes(sizeVec); });
   }
 
-  void rsSizes(std::vector<uint64_t>& sizeVec) const {
+  void rsSizes(std::vector<uint32_t>& sizeVec) const {
     sizeVec = rsFreeEntries;
   }
 
  protected:
   // Representation of the M1 Firestorm reservation station layout
-  std::vector<uint64_t> rsFreeEntries = {24, 26, 16, 12, 28, 28, 12,
+  std::vector<uint32_t> rsFreeEntries = {24, 26, 16, 12, 28, 28, 12,
                                          12, 12, 12, 36, 36, 36, 36};
   // Representation of the M1 Firestorm port layout
   const std::vector<std::vector<uint16_t>> portArrangement = {
       {0}, {1}, {2}, {3}, {4}, {5}, {6}, {7}, {8}, {9}, {10}, {11}, {12}, {13}};
   // Representation of the M1 Firestorm Reservation Station Arrangement
   // std::pair<uint8_t, uint64_t> = <rsIndex, rsSize>
-  std::vector<std::pair<uint8_t, uint64_t>> rsArrangement = {
+  std::vector<std::pair<uint16_t, uint64_t>> rsArrangement = {
       {0, 24}, {1, 26}, {2, 16}, {3, 12},  {4, 28},  {5, 28},  {6, 12},
       {7, 12}, {8, 12}, {9, 12}, {10, 36}, {11, 36}, {12, 36}, {13, 36}};
